util/requirements.txt
numpy==1.26.4
Pillow==10.3.0
wordcloud==1.9.3